Revert "Created create report API"
